requirements.txt
matplotlib==3.2.0
tensorflow_hub==0.7.0
numpy==1.16.5
tensorflow==1.15.2
pandas==0.25.1
absl_py==0.9.0
seaborn==0.9.0
Flask==1.1.1
Requests==2.23.0
gunicorn==19.5.0
scrapbook==0.2.0
numpy==1.16.5